🚚 rename: 폴더 오탈자 수정 (Jira Ticket : CCC-33)
